not so simple update
beginnings of support for dual motors over canbus, BMS, domoticz, gsm modem, radar, lights, weather, deep sleep. OTA update, tide data from uk, distance to ISS, 

dual motor telemetry works, throttle control is possible but not implemented yet, i want to implement custom traction control on my ebike to limit speed and wheel slip and power limits eventually by connecting the throttle to the screen. 

BMS supprt for Jbd Bms boards that can be found on aliexpress. i couldnt figure out how to do it over bluetooth so its using the main serial port for now, 

the domoticz home automation is used for things like weather but also for logging bike stats like state of charge and sending alerts to a phone if the vibration sensor or radar detects something.  in theory the gsm module can send data to domoticz using gprs but i have not tested that yet. 

most off the rest of the data is just pulled from online sources using json, so you can add just about anyhting you can find online to the screen, at the moment it needs to connect to the wifi on a phone to work on the move

the plan for the lights and other io is to try and get i2c working and use a gpio expander for mosfets to drive the lights and to read the switches. then a ADC for the throttle and brake.